Operations on values

SimpleSDMLayers.rescale! — Function

rescale!(layer::TI, template::TJ) where {TI <: SimpleSDMLayer, TJ <: SimpleSDMLayer}

Changes the values of the layer given as its first argument, so that it has the same range as the values of the layer given as its second argument. Modification is done in-place.

rescale!(layer::TI, t::Tuple{T,T}) where {TI <: SimpleSDMLayer, T <: Number}

Changes the values of the layer given as its first argument, so that it has the same range as the values given as a tuple of values. Modification is done in-place.

rescale!(layer::T, p::Vector{Real}) where {T <: SimpleSDMLayer}

Rescale the values of a layer so that they match with the quantiles given in p. Internally, this uses the Statistics.quantile function.

SimpleSDMLayers.rescale — Function

rescale(layer::TI, template::TJ) where {TI <: SimpleSDMLayer, TJ <: SimpleSDMLayer}

Copying version of rescale!.

rescale(layer::TI, t::Tuple{T,T}) where {TI <: SimpleSDMLayer, T <: Number}

Copying version of rescale!.

rescale(layer::T, p::Vector{Real}) where {T <: SimpleSDMLayer}

Copying version of rescale!.

Base.replace — Function

Base.replace(layer::T, old_new::Pair...) where {T <: SimpleSDMResponse}

Replaces the elements of layer according to a series of pairs. Returns a copy.

Base.replace(layer::T, old_new::Pair...) where {T <: SimpleSDMPredictor}

Replaces the elements of layer according to a series of pairs. Copies the layer as a response before.

Base.replace! — Function

Base.replace!(layer::T, old_new::Pair...) where {T <: SimpleSDMLayer}

Replaces the elements of layer according to a series of pairs. In place. Only possible for SimpleSDMResponse elements (which are mutable) and will throw an error if called on a SimpleSDMPredictor element (which is not mutable).

Missing docstring.

Missing docstring for broadcast. Check Documenter's build log for details.