Datasets

The package offers access to bioclimatic and other datasets - they are downloaded, saved to the disk, and then read locally. Please note that some of them require a lot of memory, so make sure your machine can handle them.

Worldclim 2.1

SimpleSDMLayers.worldclim — Function

worldclim(layer::Integer; resolution::Float64=10.0, left=nothing, right=nothing, bottom=nothing, top=nothing)

Download and prepare a WorldClim 2.1 bioclimatic variable, and returns it as an SimpleSDMPredictor. Layers are called by their number, from 1 to 19. The list of available layers is given in a table below.

The keywords are resolution, which must be a floating point value, and either 2.5, 5.0, or 10.0, as well as optionally left, right, bottom, and top, which will allow to only load parts of the data.

Internally, this function will download the main zip file for the required resolution from the WordlClim website, extract it, and parse the required layers.

It is recommended to keep the content of the path folder, as it will eliminate the need to download and/or extract the tiff files. For example, calling wordlclim(1:19) will download and extract everything, and future calls will be much faster.

Variable	Description

1	Annual Mean Temperature

2	Mean Diurnal Range (Mean of monthly (max temp - min temp))

3	Isothermality (BIO2/BIO7) (* 100)

4	Temperature Seasonality (standard deviation *100)

5	Max Temperature of Warmest Month

6	Min Temperature of Coldest Month

7	Temperature Annual Range (BIO5-BIO6)

8	Mean Temperature of Wettest Quarter

9	Mean Temperature of Driest Quarter

10	Mean Temperature of Warmest Quarter

11	Mean Temperature of Coldest Quarter

12	Annual Precipitation

13	Precipitation of Wettest Month

14	Precipitation of Driest Month

15	Precipitation Seasonality (Coefficient of Variation)

16	Precipitation of Wettest Quarter

17	Precipitation of Driest Quarter

18	Precipitation of Warmest Quarter

19	Precipitation of Coldest Quarter

Original data: https://www.worldclim.org/data/worldclim21.html

CHELSA V1

SimpleSDMLayers.bioclim — Function

bioclim(layer::Integer; left=nothing, right=nothing, bottom=nothing, top=nothing)

Download and prepare bioclim layers from the CHELSA database, and returns them as an array of SimpleSDMPredictors. Layers are called by their number, from 1 to 19. The list of available layers is given in a table below.

The keyword argument is path, which refers to the path where the function will look for the geotiff files.

Note that these files are large due the fine resolution of the data, and for this reason this function will return the integer version of the layers. Also note that the bioclim data are only available for the V1 of CHELSA, and are not from the V2.

It is recommended to keep the content of the path folder, as it will eliminate the need to download the tiff files (which are quite large). For example, calling bioclim(1:19) will download and everything, and future calls will be much faster.

Variable	Description

1	Annual Mean Temperature

2	Mean Diurnal Range (Mean of monthly (max temp - min temp))

3	Isothermality (BIO2/BIO7) (* 100)

4	Temperature Seasonality (standard deviation *100)

5	Max Temperature of Warmest Month

6	Min Temperature of Coldest Month

7	Temperature Annual Range (BIO5-BIO6)

8	Mean Temperature of Wettest Quarter

9	Mean Temperature of Driest Quarter

10	Mean Temperature of Warmest Quarter

11	Mean Temperature of Coldest Quarter

12	Annual Precipitation

13	Precipitation of Wettest Month

14	Precipitation of Driest Month

15	Precipitation Seasonality (Coefficient of Variation)

16	Precipitation of Wettest Quarter

17	Precipitation of Driest Quarter

18	Precipitation of Warmest Quarter

19	Precipitation of Coldest Quarter

EarthEnv landcover

SimpleSDMLayers.landcover — Function

landcover(layers::Vector{T}; full::Bool=false, path::AbstractString="assets") where {T <: Integer}

Download and prepare the EarthEnv consensus landcover data, and returns them as an array of SimpleSDMPredictors. Layers are called by their number, from 1 to

The list of available layers is given in a table below. The raw data come

from https://www.earthenv.org/landcover.

THe full keyword indicates whether the DISCover information must be included. Quoting from the reference website:

Although DISCover is based on older remote sensing imagery (1992-1993), it contains some complementary information which is useful for capturing sub-pixel land cover heterogeneity (please see the associated article for details). Therefore, it is recommended to use the full version of the consensus land cover dataset for most applications. However, the reduced version may provide an alternative for applications in regions with large land cover change in the past two decades.

It is recommended to keep the content of the path folder, as it will eliminate the need to download and/or extract the tiff files. For example, calling landcover(1:12) will download and extract everything, and future calls will be much faster. Please keep in mind that the layers can be quite large, so keeping the models stored is particularly important.

Variable	Description

1	Evergreen/Deciduous Needleleaf Trees

2	Evergreen Broadleaf Trees

3	Deciduous Broadleaf Trees

4	Mixed/Other Trees

5	Shrubs

6	Herbaceous Vegetation

7	Cultivated and Managed Vegetation

8	Regularly Flooded Vegetation

9	Urban/Built-up

10	Snow/Ice

11	Barren

12	Open Water

These data are released under a CC-BY-NC license to Tuanmu & Jetz.