These functions are not meant to be used directly, and are instead used internally by the package.

SimpleSDMLayers._match_latitude — Function

Given a layer and a latitude, returns nothing if the latitude is outside the range, or the grid index containing this latitude if it is within range

SimpleSDMLayers._match_longitude — Function

Given a layer and a longitude, returns nothing if the longitude is outside the range, or the grid index containing this longitude if it is within range

SimpleSDMLayers._layers_are_compatible — Function

_layers_are_compatible(l1::X, l2::Y) where {X <: SimpleSDMLayer, Y <: SimpleSDMLayer}

Internal function to verify if layers are compatible, i.e. have the same
size and bounding coordinates.