Reading and writing files

SimpleSDMLayers.ascii — Function

ascii(file::AbstractString, datatype::Type{T}=Float64) where {T <: Number}

Reads the content of a grid file to a SimpleSDMPredictor, the type of which is given by the datatype argument.

ascii(layer::SimpleSDMPredictor{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.

ascii(layer::SimpleSDMResponse{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.

Missing docstring.

Missing docstring for geotiff. Check Documenter's build log for details.