Simple SDM Layers

Simple SDM Layers

SimpleSDMLayers.latitudes — Function.

latitudes(p::T) where {T <: SimpleSDMLayer}

Returns an iterator with the latitudes of the SDM layer passed as its argument.

source

SimpleSDMLayers.longitudes — Function.

longitudes(p::T) where {T <: SimpleSDMLayer}

Returns an iterator with the longitudes of the SDM layer passed as its argument.

source