Datasets

The package offers access to bioclimatic and other datasets - they are downloaded, saved to the disk, and then read locally. Please note that some of them require a lot of memory, so make sure your machine can handle them.

Worldclim 2.1

Missing docstring.

Missing docstring for worldclim. Check Documenter's build log for details.

CHELSA V1

Missing docstring.

Missing docstring for bioclim. Check Documenter's build log for details.

EarthEnv landcover

Missing docstring.

Missing docstring for landcover. Check Documenter's build log for details.

ASCII files

SimpleSDMLayers.ascii — Function

ascii(file::AbstractString, datatype::Type{T}=Float64) where {T <: Number}

Reads the content of a grid file to a SimpleSDMPredictor, the type of which is given by the datatype argument.

ascii(layer::SimpleSDMPredictor{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.

ascii(layer::SimpleSDMResponse{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.