Datasets

The package offers access to bioclimatic and other datasets - they are downloaded, saved to the disk, and then read locally. Please note that some of them require a lot of memory, so make sure your machine can handle them.

By default, the layers are stored in the assets subfolder of the current project. This being said, the prefered solution is to define a SDMLAYERS_PATH environment variable pointing to a specific path, where the layers will live. This will ensure that they are re-used between projects.

All layers are returned as SimpleSDMPredictor, and therefore constructed by calling the SimpleSDMPredictor function on a LayerProvider and a LayerDataset, possibly with a future climate model and scenario. In all cases, the method accepts either a single layer, or an array of layers.

Data provider	Dataset	Layers	Future models	Future scenarios

EarthEnv	Landcover	12

EarthEnv	HabitatHeterogeneity	14

WorldClim	BioClim	19	CMIP6	SharedSocioeconomicPathway

WorldClim	Elevation	1	Elevation

CHELSA	BioClim	12	CMIP5, CMIP6	RepresentativeConcentrationPathway, SharedSocioeconomicPathway

Providers and datasets

The layernames method (inputs are a provider and a dataset) will return a tuple with the name of the layers.

Data providers

SimpleSDMLayers.LayerProvider — Type

LayerProvider

A LayerProvider is an abstract type used to dispatch the correct call of SimpleSDMPredictor to a specific dataset. A dataset is specified by a LayerProvider and a LayerDataset, as well as optionally one or multiple layers, and future climate information, resolution, or dates.

SimpleSDMLayers.WorldClim — Type

WorldClim

WorldClim offers bioclimatic data both historical and future, under CMIP6 scenarios.

This provider currently offers BioClim data, both historical and future under CMIP6, and the Elevation raster.

SimpleSDMLayers.CHELSA — Type

CHELSA

CHELSA offers high resolution climatologies. This provider currently offers BioClim data, both historical and future under CMIP5 and CMIP6.

Note that CHELSA offers a subset of all possible CMIP6 combinations, which is supposed to be the most informative.

SimpleSDMLayers.EarthEnv — Type

EarthEnv

Data from the EarthEnv project, all released under a CC-BY-NC license to Tuanmu & Jetz. This provider currently offers LandCover and HabitatHeterogeneity rasters.

Datasets

SimpleSDMLayers.LayerDataset — Type

LayerDataset

A LayerDataset is a specific set of rasters provided by a LayerProvider. For a number of dataset types that are very broad (LandCover, HabitatHeterogeneity), the precise mapping of layers is documented in their SimpleSDMPredictor method.

SimpleSDMLayers.BioClim — Type

BioClim

A list of 19 bioclimatic variables derived from the monthly temperature and precipitation data. This dataset is provided by WorldClim and CHELSA, both of which offer future versions under CMIP5 and CMIP6 models.

Variable	Description

1	Annual Mean Temperature

2	Mean Diurnal Range (Mean of monthly (max temp - min temp))

3	Isothermality (BIO2/BIO7) (* 100)

4	Temperature Seasonality (standard deviation *100)

5	Max Temperature of Warmest Month

6	Min Temperature of Coldest Month

7	Temperature Annual Range (BIO5-BIO6)

8	Mean Temperature of Wettest Quarter

9	Mean Temperature of Driest Quarter

10	Mean Temperature of Warmest Quarter

11	Mean Temperature of Coldest Quarter

12	Annual Precipitation

13	Precipitation of Wettest Month

14	Precipitation of Driest Month

15	Precipitation Seasonality (Coefficient of Variation)

16	Precipitation of Wettest Quarter

17	Precipitation of Driest Quarter

18	Precipitation of Warmest Quarter

19	Precipitation of Coldest Quarter

SimpleSDMLayers.LandCover — Type

LandCover

Information on land cover, currently only provided by EarthEnv.

SimpleSDMLayers.HabitatHeterogeneity — Type

HabitatHeterogeneity

Information on habitat heterogeneity, currently only provided by EarthEnv.

SimpleSDMLayers.Elevation — Type

Elevation

General type for a DEM, currently available through WorldClim

Future data

CMIP5

SimpleSDMLayers.CMIP5 — Type

CMIP5

Enumeration of the models from CMIP5, which can be listed with instances(CMIP5).

SimpleSDMLayers.RepresentativeConcentrationPathway — Type

RepresentativeConcentrationPathway

Enumeration of the four RCPs, which can be listed with instances(RepresentativeConcentrationPathway). These are meant to be used with CMIP5 models.

CMIP6

Missing docstring.

Missing docstring for CIMP6. Check Documenter's build log for details.

SimpleSDMLayers.SharedSocioeconomicPathway — Type

SharedSocioeconomicPathway

Enumeration of the four SSPs, which can be listed with instances(SharedSocioeconomicPathway). These are meant to be used with CMIP6 models.