Datasets

The package offers access to bioclimatic and other datasets - they are downloaded, saved to the disk, and then read locally. Please note that some of them require a lot of memory, so make sure your machine can handle them.

By default, the layers are stored in the assets subfolder of the current project. This being said, the prefered solution is to define a SDMLAYERS_PATH environment variable pointing to a specific path, where the layers will live. This will ensure that they are re-used between projects.

General interface

All layers are returned as SimpleSDMPredictor, and therefore constructed by calling the SimpleSDMPredictor function on a LayerProvider and a LayerDataset, possibly with a future climate model and scenario. In all cases, the method accepts either a single layer, or an array of layers.

Data provider	Dataset	Layers	Future models	Future scenarios

EarthEnv	Landcover	12

EarthEnv	HabitatHeterogeneity	14

[WorldClim][worldclim-current]	BioClim	19	CMIP6	SharedSocioeconomicPathway

[CHELSA][chelsa-bioclim]	BioClim	12	CMIP5	RepresentativeConcentrationPathway

[earthenv-landcover]: http://www.earthenv.org/landcover [earthenv-texture]: http://www.earthenv.org/texture [worldclim-current]: https://www.worldclim.org/data/worldclim21.html [chelsa-bioclim]: http://chelsa-climate.org/

Later providers

SimpleSDMLayers.LayerProvider — Type

LayerProvider

A LayerProvider is an abstract type used to dispatch the correct call of SimpleSDMPredictor to a specific dataset. A dataset is specified by a LayerProvider and a LayerDataset, as well as optionally one or multiple layers, and future climate information.

SimpleSDMLayers.WorldClim — Type

WorldClim

TODO WorldClim

This provider currently offers BioClim data, both historical and future under CMIP6.

SimpleSDMLayers.CHELSA — Type

CHELSA

TODO CHELSA

This provider currently offers BioClim data, both historical and future under CMIP5.

SimpleSDMLayers.EarthEnv — Type

EarthEnv

Data from the earthenv project, all released under a CC-BY-NC license to Tuanmu & Jetz. This provider currently offers LandCover and HabitatHeterogeneity rasters.

Layer datasets

SimpleSDMLayers.LayerDataset — Type

LayerDataset

A LayerDataset is a specific set of rasters provided by a LayerProvider.

SimpleSDMLayers.BioClim — Type

BioClim

A list of 19 bioclimatic variables derived from the monthly temperature and precipitation data. This dataset is provided by WorldClim and CHELSA, both of which offer future versions under CMIP5 and CMIP6 models.

Variable	Description

1	Annual Mean Temperature

2	Mean Diurnal Range (Mean of monthly (max temp - min temp))

3	Isothermality (BIO2/BIO7) (* 100)

4	Temperature Seasonality (standard deviation *100)

5	Max Temperature of Warmest Month

6	Min Temperature of Coldest Month

7	Temperature Annual Range (BIO5-BIO6)

8	Mean Temperature of Wettest Quarter

9	Mean Temperature of Driest Quarter

10	Mean Temperature of Warmest Quarter

11	Mean Temperature of Coldest Quarter

12	Annual Precipitation

13	Precipitation of Wettest Month

14	Precipitation of Driest Month

15	Precipitation Seasonality (Coefficient of Variation)

16	Precipitation of Wettest Quarter

17	Precipitation of Driest Quarter

18	Precipitation of Warmest Quarter

19	Precipitation of Coldest Quarter

SimpleSDMLayers.LandCover — Type

LandCover

Information on land cover, currently only provided by EarthEnv.

SimpleSDMLayers.HabitatHeterogeneity — Type

HabitatHeterogeneity

Information on habitat heterogeneity, currently only provided by EarthEnv.

Future climate models

SimpleSDMLayers.SharedSocioeconomicPathway — Type

SharedSocioeconomicPathway

Enumeration of the four SSPs, which can be listed with instances(SharedSocioeconomicPathway). These are meant to be used with CMIP6 models.

SimpleSDMLayers.RepresentativeConcentrationPathway — Type

RepresentativeConcentrationPathway

Enumeration of the four RCPs, which can be listed with instances(RepresentativeConcentrationPathway). These are meant to be used with CMIP5 models.

SimpleSDMLayers.CMIP5 — Type

CMIP5

Enumeration of the models from CMIP5, which can be listed with instances(CMIP5).

SimpleSDMLayers.CMIP6 — Type

CMIP6

Enumeration of the models from CMIP6, which can be listed with instances(CMIP6).

File reading and writing

SimpleSDMLayers.ascii — Function

ascii(file::AbstractString, datatype::Type{T}=Float64) where {T <: Number}

Reads the content of a grid file to a SimpleSDMPredictor, the type of which is given by the datatype argument.

ascii(layer::SimpleSDMPredictor{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.

ascii(layer::SimpleSDMResponse{T}, file::AbstractString; nodata::T=convert(T, -9999)) where {T <: Number}

Writes a layer to a grid file, with a given nodata value. The layer must store numbers.

SimpleSDMLayers.geotiff — Function

geotiff(::Type{LT}, file, bandnumber::Integer=1; left=nothing, right=nothing, bottom=nothing, top=nothing) where {LT <: SimpleSDMLayer}

The geotiff function reads a geotiff file, and returns it as a matrix of the correct type. The optional arguments left, right, bottom, and left are defining the bounding box to read from the file. This is particularly useful if you want to get a small subset from large files.

The first argument is the type of the SimpleSDMLayer to be returned.

geotiff(file::AbstractString, layer::SimpleSDMPredictor{T}; nodata::T=convert(T, -9999)) where {T <: Number}

Write a single layer to a file, where the nodata field is set to an arbitrary value.

geotiff(file::AbstractString, layers::Vector{SimpleSDMPredictor{T}}; nodata::T=convert(T, -9999)) where {T <: Number}

Stores a series of layers in a file, where every layer in a band. See geotiff for other options.

geotiff(file::AbstractString, layer::SimpleSDMResponse{T}; nodata::T=convert(T, -9999)) where {T <: Number}

Write a single SimpleSDMResponse layer to a file.

geotiff(file::AbstractString, layers::Vector{SimpleSDMResponse{T}}; nodata::T=convert(T, -9999)) where {T <: Number}

Write a vector of SimpleSDMResponse layers to bands in a file.