Other operations

Other operations

Grid coarsening

SimpleSDMLayers.coarsen — Function.

coarsen(L::LT, f::FT, d::Tuple{IT,IT}; NaNremove::Bool=true) where {LT <: SimpleSDMLayer, FT <: Function, IT <: Integer}

This function will aggregate a layer by merging cells. The function f passed as its second argument is expected to take an array as input, and return a single value of the same type as the values of the layer, or as a floating point value. Note that the element type of the returned layer will have type Any, which is not really pretty, but should serve well enough (and can be changed by working on the grid field directly if you really need it).

The size of the cells to aggregate is given by the tuple, so that (2,2) will make groups of two cells vertically and two cells horizontally, for a total of four cells.

The NaNremove keyword argument is intended to remove NaN values before applying f. It defaults to true.

source