Overloads

Base methods overloaded

To facilitate writing julian code, we have overloaded a number of methods from Base. These methods should remove the need to interact with the grid field directly, and also allow to set and get values using the geographic coordinates (as opposed to the grid positions).

Base.convert — Function.

Base.convert(::Type{SimpleSDMResponse}, p::T) where {T <: SimpleSDMPredictor}

Returns a response with the same grid and bounding box as the predictor.

source

Base.convert(::Type{SimpleSDMPredictor}, p::T) where {T <: SimpleSDMResponse}

Returns a predictor with the same grid and bounding box as the response.

source

Base.eltype — Function.

Base.eltype(p::T) where {T <: SimpleSDMLayer}

Returns the type of the values stored in the grid.

source

Missing docstring.

Missing docstring for size. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eachindex. Check Documenter's build log for details.

Base.getindex — Function.

Extracts a value from a layer by its grid position.

source

Extracts a series of positions in a layer, and returns a layer corresponding to the result. This is essentially a way to rapidly crop a layer to a given subset of its extent. The i and j arguments are UnitRanges (of Integer).

The layer returned by this function will have the same type as the layer passed as its argument.

source

Extracts the value of a layer at a given latitude and longitude. If values outside the range are requested, will return NaN.

source

Extracts a series of positions in a layer, and returns a layer corresponding to the result. This is essentially a way to rapidly crop a layer to a given subset of its extent. The longitudes and latitudes arguments are tuples of floating point values, representing the bounding box of the layer to extract.

The layer returned by this function will have the same type as the layer passed as its argument.

source

Extract a layer based on a second layer – because the two layers do not necesarily have the same origins and/or strides, this does not ensure that they can be overlaid, but this is a good first order approximation.

source

Missing docstring.

Missing docstring for setindex!. Check Documenter's build log for details.

Base.similar — Function.

Always returns a SimpleSDMResponse

Returns NaN for NaN, and eltype zero for other values

source

Base.copy — Function.

Returns the same type

source