Methods overloaded

To facilitate writing julian code, we have overloaded a number of methods from Base. These methods should remove the need to interact with the grid field directly, and also allow to set and get values using the geographic coordinates (as opposed to the grid positions).

From Base

Base.convert — Function

Base.convert(::Type{SimpleSDMResponse}, p::T) where {T <: SimpleSDMPredictor}

Returns a response with the same grid and bounding box as the predictor.

Base.convert(::Type{SimpleSDMPredictor}, p::T) where {T <: SimpleSDMResponse}

Returns a predictor with the same grid and bounding box as the response.

Base.convert(::Type{Matrix}, p::T) where {T <: SimpleSDMLayer}

Returns the grid as an array.

Base.copy — Function

Base.copy(l::T) where {T <: SimpleSDMLayer}

Returns a new copy of the layer, which has the same type.

Base.eltype — Function

Base.eltype(p::T) where {T <: SimpleSDMLayer}

Returns the type of the values stored in the grid.

Base.size — Function

Base.size(p::T) where {T <: SimpleSDMLayer}

Returns the size of the grid.

Base.stride — Function

Base.stride(p::T; dims::Union{Nothing,Integer}=nothing) where {T <: SimpleSDMLayer}

Returns the stride, i.e. the length, of cell dimensions, possibly alongside a side of the grid.

Base.eachindex — Function

Base.eachindex(p::T) where {T <: SimpleSDMLayer}

Returns the index of the grid.

Base.getindex — Function

Extracts a value from a layer by its grid position.

Base.getindex(p::T, i::R, j::R) where {T <: SimpleSDMLayer, R <: UnitRange}

Extracts a series of positions in a layer, and returns a layer corresponding to the result. This is essentially a way to rapidly crop a layer to a given subset of its extent. The i and j arguments are UnitRanges (of Integer).

The layer returned by this function will have the same type as the layer passed as its argument, but this can be changed using convert. Note that this function performs additional checks to ensure that the range is not empty, and to also ensure that it does not overflows from the size of the layer.

Base.getindex(p::T, longitude::K, latitude::K) where {T <: SimpleSDMLayer, K <: AbstractFloat}

Extracts the value of a layer at a given latitude and longitude. If values outside the range are requested, will return NaN.

Base.getindex(p::T; left::K=nothing, right::K=nothing, top::K=nothing, bottom::K=nothing) where {T <: SimpleSDMLayer, K <: Union{Nothing,AbstractFloat}}

Returns a subset of the argument layer, where the new limits are given by left, right, top, and bottom. Up to three of these can be omitted, and if so these limits will not be affected.

Base.getindex(p1::T1, p2::T2) where {T1 <: SimpleSDMLayer, T2 <: SimpleSDMLayer}

Extract a layer based on a second layer. Note that the two layers must be compatible, which is to say they must have the same bounding box and grid size.

Base.getindex(p::T, occurrence::GBIF.GBIFRecord) where {T <: SimpleSDMLayer}

Extracts the value of a layer at a given position for a GBIFRecord. If the GBIFRecord has no latitude or longitude, this will return NaN.

Base.getindex(p::T, r::GBIF.GBIFRecords) where {T <: SimpleSDMLayer}

Returns the values of a layer at all occurrences in a GBIFRecords collection.

Base.setindex! — Function

 Base.setindex!(p::SimpleSDMResponse{T}, v::T, i...) where {T}

Changes the value of a cell, or a range of cells, as indicated by their grid positions.

Base.setindex!(p::T, v, lon::Float64, lat::Float64) where {T <: SimpleSDMResponse}

Changes the values of the cell including the point at the requested latitude and longitude.

Base.setindex!(p::T, v, occurrence::GBIFRecord) where {T <: SimpleSDMResponse}

Changes the values of the cell including the point at the requested latitude and longitude.

Base.similar — Function

Base.similar(l::T) where {T <: SimpleSDMLayer}

Returns a SimpleSDMResponse of the same dimensions as the original layer, with NaN in the same positions. The rest of the values are replaced by the output of zero(eltype(p.grid)), which implies that there must be a way to get a zero for the type. If not, the same result can always be achieved through the use of copy, manual update, and convert.

Missing docstring.

Missing docstring for Base.sum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.maximum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.minimum. Check Documenter's build log for details.

From Statistics

Missing docstring.

Missing docstring for Statistics.mean. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Statistics.median. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Statistics.std. Check Documenter's build log for details.