Types

Types

Layers are represented by a grid, storing the content of cells in a Matrix, and a bounding box indicated by the floating point coordinates of its limits.

Abstract type

SimpleSDMLayers.SimpleSDMLayer — Type.

All types in the package are part of the abstract type SimpleSDMLayer. A SimpleSDMLayer has five core fields: grid is a matrix storing the cells, and left, right, bottom and top are floating point numbers specifying the bounding box.

It is assumed that the missing values will be represented as NaN, so the "natural" type for the values of grid are floating points, but it is possible to use any other type T by having grid contain Union{T,Float64} (for example).

source

Implemented types

SimpleSDMLayers.SimpleSDMResponse — Type.

A response is a SimpleSDMLayer that is mutable, and is the usual type to store analysis outputs. You can transform a response into a predictor using convert.

source

SimpleSDMLayers.SimpleSDMPredictor — Type.

A predictor is a SimpleSDMLayer that is immutable, and so does not have methods for setindex!, etc. It is a safe way to store values that should not be modified by the analysis. Note that if you are in a bind, the values of the grid field are not immutable, but don't tell anyone we told you. The correct way of handling predictors you need to modify would be to use convert methods.

source